Some explicit typecasts; they don't look like they're necessary, but it doesn't compile on Linux/Alpha without them (and they can't hurt).
svn path=/trunk/kdelibs/; revision=46637

+++ khtml/rendering/render_table.cpp
@@ -1245,7 +1245,7 @@ void RenderTable::layoutRow(int r)
 
     	if (style()->direction()==RTL)
     	{
-	    cell->setPos( columnPos[totalCols]-columnPos[indx+cell->colSpan()],
+	    cell->setPos( columnPos[(int)totalCols]-columnPos[(int)(indx+cell->colSpan())],
 	    	rowHeights[rindx] );
 	}
 	else
